+++ src/renderer/src/api/browser-api.ts
@@ -627,12 +627,14 @@ export class BrowserApi implements ApplicationAPI {
     });
   }
 
-  runCodeInlineRequest(baseDir: string, filename: string, lineNumber: number, userComment: string): void {
+  runCodeInlineRequest(baseDir: string, taskId: string, filename: string, lineNumber: number, userComment: string, createNewTask?: boolean): void {
     this.post('/project/run-code-inline-request', {
       projectDir: baseDir,
+      taskId,
       filename,
       lineNumber,
       userComment,
+      createNewTask,
     });
   }